Subscribe to a socket on a Transmitter
This will allow us to send the socket to the processes that need to
publish QoS0 messages to the broker. No need to pass the data to the
transmitter for that.

When the subscribing process terminate the transmitter will clean up
its subscriber state; when the transmitter receive a new socket it
will broadcast the socket to its subscribers.