Changed the publish and connect structs to allow nil for not set
It is now possible to specify the last will of a connect as nil for
not set. Previously one had to pass in a publish with payload set to
nil, but this is not correct as it is allowed to have a publish with
an empty payload.

Had to change the eqc models a bit to reflect this. I think they are
more correct now, and they still pass.