Get rid of connect and reconnect functions in the Connection module
Previously there was a legacy function called connect that was mostly
used from the unit tests; the actual connection logic was in the init
function of the Connection gen_server implementation.

Furthermore there was a reconnect function that told the connection to
reconnect, using a slightly different start up strategy; it does not
need to start the supervision tree as it has already been started in
the init.

All this has been replaced by having the connection process monitor
the server socket, and run the reconnect on the down message. This
simplifies the receiver process even more as we do not need to send a
message from the receiver to the connection to reconnect; and there
are fewer public function on the connection.