Fixed a blinking test that was caused by a race condition
Sometimes the socket was closed before it was handed from the
connection module to the receiver module. This would cause the
handover call to fail as the socket was closed.

So far I have just added a match for that case and return :ok, but we
should probable give this some thought