Move the generation of trackers into the inflight process
This will allow us to assign an unassigned identifier to the packages,
so we can have the erlang refs be the interface of uniqueness when we
face the user, and the 2-byte identifier the uniqueness when we face
the server. By having the inflight process handle this it can be the
authority on what identifiers are in play at the given time.