Ping requests will now return a reference and reply back when done
When issuing a ping the calling Pid will now get a reference, and when
the ping is done it will get a message in its mailbox:

    {Tortoise, {:ping_response, ^ref}}

This is done so we can have the connection ping the server once in a
while so we will adhere to the "keep alive" and this will allow us to
kill the connection if the response does not come within "reasonable
time"; reasonable time should probably be a configuration setting...

This implementation should also allow for multiple ping requests at
the same time. This does not make much sense but it is possible.