Subscribe to topics based on a list given to the connection on init
It is now possible to give the connection configuration a list of
topics the client should connect to when it has a connection. For now
I have decided to let it subscribe when it opens its first connection
and when it reconnect with a clean state. I have to look into if the
subscription state is kept if there is a previous state on the server
when the client connect with a non clean state.