The Connection will now ensure we send pings to fulfill keep alive
The Connection module will use the keep alive information in the
Connect to send ping requests to the server on the given keep alive
interval. By placing the responsibility there we can setup integration
test scenarios where we do not have to care about ping/ping responses
in the scenario.

The only downside is that we will send the keep alive message even if
there has been activity, but this is really hard to figure out anyways
because we support handing the sockets for writing to other processes,
so there are no central authority in the system that can tell if there
has been outgoing activity or not.

Sending ping requests according to the keep alive is fine anyways as
we can use the data to see how long the round-trip to the server is
over time.