Retrieve a new pipe if the current socket is down during a publish
When a process holds a pipe containing a socket that is down it will
enter a selective receive when it attempt to publish on this
socket. The selective receive will listen for the new socket being
send to its mailbox by the transmitter, and use this new pipe to send
the message.

If the receive does not receive the new pipe within five seconds it
will return an error-tuple. Perhaps the timeout should be a
configuration setting on the pipe.