Remove the `ping_response/2` callback from the Driver
Ping requests and responses are used to keep the connection between
the client and server alive. When the client connect to the server a
keep alive interval is specified; if the server does not see a message
in this interval it will close the connection--so the ping requests
are used to keep the connection alive, which is a protocol detail the
user should not concern themselves with.

We time the time it takes for the ping request is sent to the ping
response is received, which allow us to time the package round
trip. Previously this information was made available to the
driver. This is no longer available, but if the client request a ping
the response will contain the round trip time in μs.