Implement `subscription/3` callback in the driver callback module
When a topic filter changes state (is subscribed to or is unsubscribed
from) the callback `subscription/3` is triggered in the user
implemented driver module.

I had to send a message from the inflight process to the controller
process, as the result of a tracked subscribe or unsubscribe package
is needed for the user callback. I will consider if the "report back
to caller" should reside in the Controller process as well. It kind of
make sense to put it there. Then the Inflight process would only be in
charge of progressing a tracked package to completion and hand it off
to the controller which will decide on what to do with the result.

Some open questions are left; what should happen to subscriptions
where a different quality of services is returned than then one that
was requested?