Moved respond caller from the inflight process to the controller
Some of the calls does not go to the process but are still executed in
the inflight process. I think this makes the inflight module prettier
and the reporting back to the controller which before seemed like a
hack seems like design...at least I like it better this way.