Implement the `connection/2` driver callback
This will trigger when the connection changes status; it can either be
`:up` or `:down`; the callback will only change when it flips from one
to the other.

I implemented a default-driver as well. This one does nothing to the
data passed through and can be used in unit tests where there the
callback module is not tested.