Ignore the reported QoS in the subscription callback
The system will report the accepted QoS back when a subscription has
been acknowledged; if we request a subscription of QoS 2 we can have
the subscription acknowledged with a QoS of 1. This depend on the
server configuration.

I think we should consider crashing if a non-requested QoS is
accepted. That way the user can fix their server, or otherwise work
around the problem. Somehow we have to deal with this problem, because
the user could expect a certain behavior, and see something that
seemingly works, but in reality works differently than expected.

We need to consider how to deal with this.