Implement `ping_sync/2` on the Controller
Also switched to using the new `ping_sync/2` on the
`Tortoise.Connection.Controller` instead of the one implemented
directly on the `Tortoise.Connection`-module.

If we change the implementation of the ping await the Controller will
get the change as well. Also this provides a `ping_sync/2` which can
be used by the user to ping the server without having to pull out the
response from the mailbox themselves.

The first argument is the `client_id` the second one is the timeout
given in milliseconds.