Restructured the transmitter and pipe modules a bit
The Transmitter module knows nothing about "pipes" now, it only deals
with sockets and processes who request them and subscribe to them. It
can cast data onto the wire, but that should be about it.

The pipe module knows how to put data on a TCP connection, and the
publish function has been moved there. I have implemented passive and
active subscription modes. They are to be though of as:

  - "Active" will make a subscription on the transmitter, every time
    the transmitter receive a new socket it will rely the new socket
    to the processes holding pipes with active subscriptions.

  - "Passive" will not get listed as a subscriber on the transmitter,
    but will ask for a socket if the one it has return an error such
    as `{:error, :closed}`

The active subscription will require more work for the user as a
handle_info needs to be set in place that will update the pipe kept in
the state; the passive one might have a slight delay when it needs to
renew its socket, but it will all happen behinds the scenes. It will
be interesting to see how this works out when we start implementing
pipe strategies; perhaps the "active"-kind is not needed at all;
perhaps it is in cases where we want short timeouts, and dropping
packages are okay until we get a new socket from the transmitter (but
waiting 5 seconds for the socket is unacceptable because packages
might build up)