Start implementing subscribe on the Tortoise.Connection
For now I just moved the subscribe function to the
Connection-module. For testing I needed to be able to send in an
identifier so I can send back the correct suback message in my tests,
and for now it is making the call in the calling process. This need to
be moved into the Connection process so we can update the subscription
list when there are made changes to it.