Handle subscription via the Connection process
Changed the return of the tracking of a subscribe package generated
when it is finalized in the tracker. Now it will always return a map
containing the keys: `:ok`, `:warn`, and `:error`. This makes it
easier to make pattern matches.

There still need to be some error handling. Right now it will ignore
warns (it should not) and it will just stop on straight up errors; I
think this should be a configuration option. Crashing would be
sensible if we cannot subscribe to a topic, because the requirements
to the broker was not met.