Rename driver to handler
The name of the user implemented callback that is used by the
connection to decide on what to do when a message is received from the
broker, or a subscription made or removed, or when the connection is
lost and we reconnect again is now called a Handler.

I think Handler is a much better term for this, because Driver might
be confused with a thing that handle the network communication.