Attempt to make the dialyzer pass
I had to remove the spec from the `Track.create/2` function; at this
point I cannot see what it is complaining about.

I also had to upgrade to the master on github for the GenStateMachine
dependency, as it seems that version fix a dialyzer issue that I was
hitting. This version has not been published on hex yet. I have talked
to the maintainer and hopefully he will get around to making a new
release in the near future.

This is far from perfect. I need to work a bit with the specs, perhaps
if I guide dialyzer better by adding more specs to the public
functions it would have a better time.