Change `publish_sync/5` to `publish_sync/4` on the Tortoise module
Previously the function accepted a timeout, an integer value (timeout
in milliseconds) or an atom of `:infinity`, as the last argument. This
has been moved to the options list passed in. This allow us to take
advantage of the syntactic sugar Elixir provides when the last
argument is an options list.

Previously:

    Tortoise.publish_sync(
        "client_id",
        "foo/bar",
        "hello",
        [retain: true, qos: 1],
        timeout: 200
    )

After this change:

    Tortoise.publish_sync(
        "client_id",
        "foo/bar",
        "hello",
        retain: true, qos: 1, timeout: 200
    )