Implemented an SSL transport
Added the ct_helper from ninenines allowing us to create a listener
that works with the SSL transport. Added a test the make sure we can
connect using a scripted MQTT server that is running in SSL mode.

Added :ssl as an extra_application to the mix file so the dialyzer
wont complain about the unknown ssl application.

We can no longer monitor the network socket in the connection process
as the SSL module will use an opaque data structure instead of a
socket that can be used in `Port.monitor`; instead we send a
"renew"-message from the receiver process to the connector when the
network goes down.