Change the connection spec
When specifying the connection one would previously specify the
connection as `{:gen_tcp, host, port}`. Now that we have multiple
transport types it is better to change that to the format of
`{transport, opts}`, where the transport is one of

  - Tortoise.Transport.Tcp
  - Tortoise.Transport.SSL

The opts is a keyword list, allowing the individual transport to
accept the options relevant to it.

- Factorize the creation of the connection spec into
  Tortoise.Transport.new/2, that will then dispatch to the given
  transport, allowing us to error with a good error message if the
  wrong opts has been passed to the transport.

- Update changelog

- Bump to version 0.2.0