Implement ok-3 tuple for the handle_message handler callback
Add a next_actions to the Handler struct, and accept a ok-3 tuple with
the form of `{:ok, new_state, actions_next_list}` as a return value
from then handle_message callback. The handling of the return is
lifted to a helper that will make it easier to support the returns for
all the callbacks. Again, so far it is only implemented for
handle_message, and the nothing is done with to the next actions list.