Make the subscription (down) handler callback accept 3-ok tuples
This is done by reducing the unsubacked topics, one by one, into the
handler state. This will allow us to extend the next actions list if
one of more of the unsubbed topics creates work on the next actions
list, and execute that work when the callbacks has been called for
each of them.