Made an async version of subscribe
This is needed for the next_actions, as they cannot block the
controller. This require us to handle async results as handle_info in
the controller process; I've implemented it such that the controller
process knows what is in flight, and will error if it get a reference
to something unknown back.

The old `subscribe/3` has been renamed to `subscribe_sync/3` and has
the same behavior.