Further documentation of the `Tortoise.Connection` module
I made the `renew/1` connection "public" and renamed it to
`reconnect/1`; this is done so it can be documented and an operator
can use it to reconnect to a broker, or a device could have a button
that will use this function in part of a soft reset.

Besides that I added more words.