Return an error on `{:error, :nxdomain}` when connecting
This is not the intended behavior; we need to make it reattempt the
connection a couple (or infinite) times, with an exponential back-off,
according to a configuration option.

This introduce a GenServer backed `Tortoise.Transport` that will allow
us to setup error scenarios, with specific errors, such as
`{:error, :nxdomain}`.