Move the initial connect outside of the Tortoise.Connection `init/1`
This will allow for a connect to take a bit more time than the
supervisor would allow; the tortoise connection should no longer block
its supervisor until it have a connection. This changes the ways the
connect can fail though. Good thing is that before there was two ways,
now we have to potential of having one.

The update here might call for some refactoring. Perhaps we can
combine the connect and reconnect. The error reason and blame should
be split; when the connection fail because of an error in the
configuration (no cacert file specified for instance) there would be
no reason the retry the connect.