Store the network socket in ets and move publish to Tortoise
The publish functions has been moved to the `Tortoise`-module. It is
no longer possible to publish using `Tortoise.Connection.publish/4`,
instead one should just use the `Tortoise.publish/4`, which previously
defdelegated to `Tortoise.Connection`.

The connection will store the socket in an ets table in the
Tortoise.Registry; this will allow us to answer the question: is the
connection even up? Previously we would just time out if a message was
sent using a non-existent connection.