Make it possible to await a socket from a connection
This was previously possible through the transmitter process, but the
functionality has been moved to the Tortoise.Registry and the
Tortoise.Connection. The publish functions has been updated to use
this instead, and the pipe struct will fetch a connection from the
connection/registry as well.

Hopefully the Registry will get an await feature in the future that
can be used for these purposes. If that happens this functionality
should get moved to that.