Add a Tortoise.Events registry
Replace the connection subscriptions from the connection
module/process with a Registry based one. This allow us to test the
subscription without having to set up most of the system. This will
also allow us to make a GenEvent like system where events such as a
disconnection can be subscribed to.