Get rid of the Tortoise.Connection.Transmitter module
Tortoise had a state machine that kept the current connection and a
buffer of packets to transmit. After we started to send the network
socket to the process that need to put data on the wire the
transmitter took a lesser role; it was only called by the Inflight
process and for sending ping requests in the controller.

We have now switched to requesting a network socket for the ping
requests, and the inflight process will ask for one as well.

Also, for some reason Tortoise had support for answering ping requests
sent from the server; this is not possible according to both the MQTT
3.1.1 and the MQTT 5 specifications, so it has been removed and will
instead stop the connection with a protocol violation reason.