Updated track data structure and ported inflight to a gen_statem
The new track data structure is better in the sense that it has a left
side, which is an action we need to take, and a right side, which is
the expected response to the action on the left side. When we receive
something that makes the stage move forward we will never need to go
back; and because we only move forward when we get the expected
response we can easier implement a retry functionality; at this point
we just need to retry the current pending action.

It passes the tests, but currently it doesn't clean up after it self;
I will address this in a coming commit.