Implement a `Tortoise.Connection.disconnect/1`-function
The `disconnect/1` function will cleanly terminate a connection;
meaning it will send a cancel message to the processes awaiting a
response to messages in-flight, and it will send a disconnect message
on the wire, after which it will close the network socket.

The inflight manager will set its status to draining which will allow
us to ignore messages in the mailbox; should a process attempt to
publish a message with QoS>0 after the inflight monitor have switched
its state to draining it will get the error message `{:error,
:terminating}`.

There is a slight chance that another process could put a message on
the socket after the disconnect-package and before we get to close the
socket. According to the specification it is not allowed to have more
data on the socket after the disconnect, but I don't think there is
much we can do about this. This is a theoretical issue though. I don't
think one should worry too much about it.

Notice; for now we will cancel outbound messages, perhaps we could
drain the outbound messages by making sure they are delivered, while
not accepting any new outbound messages. Starting off it seems easier
to just cancel the messages though.