Register for connection before requesting it to avoid race condition
There was a small window when the process would register for a
connection after it had gotten a the `:connecting` status from the
registry; in this window the connection could be come available, and
the pubsub would dispatch the connection before the registration
happened, in which case it would reach the receive block too late.

By setting up the connection subscription before checking the registry
we can avoid this problem. We will then unregister the subscription
after the function is done. The down side is that we will need to hit
the ETS table every time we get a connection, but at least we will
avoid processes timing out or even hanging forever when a connection
is requested.