WIP restructure of the entire application to support MQTT 5
The controller has been merged with the connection process as we now
need to support auth packages; thus we cannot just receive the fixed
number of bytes when we await the connack message. The receiver will
have to be booted earlier in the process, so that has been rewritten.

The connection process is now a gen_statem; and it will handle the
user specified callback module in the future.

This is far from done. Some of the protocol logic has been
implemented, but it is far from a MQTT client right now.