Make sure the receiver is up before trying to connect
This is put in place as the receiver would reconnect too slowly after
a crash; the connection handler would attempt to reconnect before a
new receiver was ready.

The solution is to await a ready message from the receiver, which will
get dispatched when the receiver is in its init function. This is only
done if we do not have a reference to an existing receiver.

This could perhaps be solved more elegantly if the Elixir Registry had
some kind of "await" function like gproc does. If Registry get an
await function we should replace this with it.