Start implementing disconnect messages and handle_disconnect
Now I will start to change the user facing callback module a bit. The
old style of handler was perhaps a bit too simple; MQTT 5 require us
to make the actions returned from the callbacks a bit more advanced;
for instance when we handle a disconnect we might want to reconnect,
or stop the client--at the end the user need to make this decision
because there are no one size fits all solution to this.

The execute function in the handler will get split into multiple
functions; in future commits we will see other states such as
handle_auth and handle_connack. They will map into a state in a state
machine, and the possible options will depend on where we are in said
state machine.

Unfortunately this will force the user to know more about the
internals of MQTT, but with some good sane defaults for the states we
should still be able to make an easy to use interface.