Re-do the ping/keep-alive logic
The old version seemed to leak because the keep alive messages created
references in the ping queue. The new implementation will keep a
timeout for the keep alive message, and dispatch a ping once it is
triggered. If an outside process request a ping we will cancel the
keep alive timeout reference and dispatch a ping. If another outside
process request a ping while the ping has been dispatched to the
server it will latch on to the result and respond to the process with
that value.

This allow us to get rid of the ping-queue; which admittedly was a bit
over-engineered--and as the keep alive is reset after we receive the
ping response we will ping the server less.