Add and expose a handle_pubrel callback
I have no idea what this should be used for, but the pubrel package
can contain a user defined property, so better expose it to the
end-user. Also, the end user need to be able to define user defined
properties on the pubcomp; that will come in a later commit.