Support user defined properties for protocol packages
The inflight process will now await a package to dispatch before
progressing tracking states. Before it just dispatched the next
message in the plan, but that is not possible anymore as we need to
support user defined properties on packages such as pubrel, pubrec,
pubcomp, etc.

The controller has been updated so it will pass the dispatch update
into the inflight process; the tests has been updated as well.