Removed the subscriptions field from the internal state
The :subscriptions field on the internal connection state was used to
keep a list of user defined subscriptions for a connection. My thought
was to move it to a configuration object, so it would be outside of
the state and thus be able to survive a crash, and have the connection
connect with the same subscription for a given client id, but this is
impossible because of user defined properties!

We cannot just pack all subscriptions in one subscribe call, as the
user might want to set a user defined property for one of the
subscriptions, and we cannot just set the user defined property in the
configs along with the subscription as the user might want to send new
data every time they subscribe.

User defined properties sure are a blessing and a curse.