Make it possible to transform packages in user initialized publishes
The publishes now take a list of transform functions, operating on the
package types send and received in the publish exchange. This makes it
possible to intercept the pubrel and add user defined properties on
them, specified for the individual publish. It has a transaction state
as well, so state can be carried from publish, pubrec, pubrel, and
pubcomp, etc.