Remove the need for a magic pid variable in the callback TestHander
I might have replaced it with some other magic, but for most of the
tests we really did not need to send back the input, as it for the
most part passed in untouched. It is possible to send in a function
that send values back to the calling process, so I have changed it to
just doing that. Also, I have made it possible to send in a function
for the execute_init, and I have added more tests for the init
callback.