Introduce a eval-next action
It is now possible to execute an anonymous function in the context of
the tortoise connection by passing it in as a next action with the
form `{:eval, fn (state) -> IO.inspect(state) end}`.

The `state` given to the function is the connection state. It is not
possible to alter the state, as the return of the function is
disregarded.

This has been added to make it easier to test execution of next
actions in the callback tests.