Change the format of user properties
It is no longer possible to specify a 2-tuple of `{utf8, uft8}` when
defining user properties. Now a user property has to be specified as
`{:user_property, {utf8, utf8}}`. This might not be the most elegant
solution, but it allow us to preserve the order of the properties in
the encoded package.