Renamed the connection/2 to status_change/2
The status change event is meant to update when the client is online
and when it is dropped from the broker. There is also a "terminating"
phrase, so the user can act on this stage.

I liked the name status_change better, and it is now an optional
callback; if it is not implemented Tortoise will just skip this and
continue the life-cycle unchanged.