Allow a topic alias instead of a topic in Tortoise.publish(_sync)/4
When a topic alias is given instead of a proper topic the topic will
be set to an empty string and the given topic alias will be set in the
property list as `{:topic_alias, topic_alias}`. Tortoise will not
check if the topic alias is set, as we have no way of knowing.