Made it possible to specify properties and reasons on disconnect
The `disconnect/1` on `Tortoise.Connection` has been extended and is
now `Tortoise.Connection.disconnect/3`. Second argument is the
disconnect reason, and the third is the properties. Both has defaults
and respectively default to `:normal_disconnection` and the empty list
`[]`.

We do not check the disconnect `reason` effectively allowing the user
to specify a disconnect normally only allowed by the server, but some
brokers might do crazy stuff with this, so let us just allow the user
to do that if they want.