Update the client subscription state when unsubscribing from topics
It is possible to ask a connection for what connections it is
currently connected to. When we posted an unsubscribe package to the
server we would remove the unsubscribed topics from this internal
subscription list. Now we will only remove the topics that were
successful (or wasn't a subscription in the first place). This handle
situations where we are not allowed to unsubscribe from a topic; in
this case the client would still report that it is subscribed to the
topic.