Store the server set keep alive interval if present in connection
If the server specify a keep alive in the connack it should get used
for the keep live interval. We will not pick the server set value if
present, and keep the user defined if not.

Also, this commit introduce a `Tortoise.Connection.info/1` that will
return the connection configuration when the client is connected.