Simplified merging of properties into connection config
Now that we cast package property values into booleans it is much
easier to merge the connection config; actually it might be overkill
to check if the values exist in the config object.

Actually, this might pose a problem if the connack contain user
defined properties; but let's worry about that later.