Handle empty subscription topic filter list by sending err to caller
The user facing API has a check for an empty topic filter list in a
subscription, but for good measure, and to remove an unless-expression
I have made a special case for subscribe casts with an empty topic
filter list. If the caller is external we will send a error empty
topic filter list error back.