Don't allow shared subscriptions if the server doesn't allow them
Introducing a validator step when placing a subscription. For now it
will check if there are any shared topic filters in the subscription
topic filter list when placing a subscription on a server that is
configured to not allow subscriptions to shared topic filters. If this
is done it will result in an error send to the user.