Rename config to info, and store capabilities under capabilities
Previously I stored the server capabilities under a key called
config. I fear that that would have led to confusion as the user would
expect to be able to change this config. In reality it is server
capabilities, so I called it that instead. Besides that I moved the
capabilities into a struct containing connection "info." This will
allow us to respond with more data when the user request info about
the connection, as we can store the client_id, keep_alive interval,
etc there.