First step in getting rid of the Tortoise.Events
When the network socket is dropped we need reconnect and get a new
one; the new network socket needed to get communicated to interested
parties, such as the inflight manager. This relied on meta data in the
tortoise registry, and the meta data rely on the client_id. We need to
get away from using the client id as the instance identity, so this is
the first step in getting out of that.

Also, we need to let the user control the connection flow, so they can
add their own backoff strategies and decide on when tortoise should
connect. These changes will hopefully make this easier to pull
off. got rid of some seemingly dead code as well.