Implemented a generator for connack messages
Added a protocol for defining structs that can be generated; I
implemented the generatable protocol for Connack messages, as they
seem fairly simple, yet not too simple. Had to figure out how to
generate properties, and it seems that my solution can error out at
times because it cannot make the list unique enough (we only allow
keys to occur once in the property list, except for the user defined
properties).

This should only load if StreamData is included in the project so we
can make StreamData an optional dependency, and hopefully this
approach can be used to test more than just Tortoise.

When generating the data one has to set the fields they want to
generate to `nil`; if they are set to any other value the generator
will pass the set value through as a constant. This allow us to only
generate successful packages like so:

```elixir
iex(1)> alias Tortoise.Package
Tortoise.Package
iex(2)> %Package.Connack{reason: :success, properties: nil} |> Package.generate() |> Enum.take(1)
[
  %Tortoise.Package.Connack{
    __META__: %Tortoise.Package.Meta{flags: 0, opcode: 2},
    properties: [maximum_qos: 0],
    reason: :success,
    session_present: false
  }
]
```

Connack encoding and decoding is now tested using this generator.