Add a data generator for suback messages and a prop test for them
When generating the ack list it is possible to specify `nil` which
will generate a list of acks; but it is also possible to steer the
generation such that `[nil]` will generate a list with one random ack,
and `[nil, nil]` will generate a list with two random acks. One could
also specify `[{:ok, nil}, {:error, nil}]` which will generate a list
of length two with an random ok-ack on the first position and a random
error reason on the second.