Add a data generator for Publish packets and test them
I needed a way to update the flags in the fixed packet header, and
these values had to be inferred from the content of the package, so I
added a `infer/1` function on the Meta module, which will ensure the
fixed header is correct before encoding.

After introducing this some of the tests started failing because the
fixed header was wrong; this has now been fixed.

I still need to make the user defined properties generate other
properties than the user defined ones.