Replace the inflight system with an ETS based one
This is part of the work to get rid of the client id as the identifier
for the connection. The inflight messages are now stored in a ETS
table; outside of the connection; so it can survive a connection
rolling over, and eventually it would support loading a session into
memory and start from that.

It is not 100% complete.

Also, it has a notion of backends, and the ETS is default, but it
would be possible to back it with external services such as an SQL
server or a Redis, which would allow us to tear down an entire node
and continue on another.