Restructure the transmitter supervision for the connections
Introducing a transmitter supervisor that is managed by the tortoise
application; when tortoise start a connection it will start a
connection using this (dynamic) supervisor, and both the transmitter
and the connection will monitor each other. Previously the connection
would spawn link a supervisor that held the connection, and that was
because we used to store more data in multiple processes; but after
the session supervisor has been introduced we only kept track of the
transmitter in the connection supervisor, and at this point we might
as well simplify things and pull it out. Because it is monitored it
should clean up as expected when one of them terminates.